backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
pydantic>=2.10
python-dotenv>=1.0
openai>=1.40
openpyxl>=3.1
python-multipart>=0.0.20
